interview-service-ai/InterviewReportService/requirements.txt
kafka-python==2.2.15
python-dotenv==1.1.1
mongoengine==0.29.1
redis==7.1.0
google-generativeai==0.8.5